+++ R/preamble.R
@@ -3,8 +3,11 @@ the$defaultcachedir <- tools::R_user_dir("alphavantagepf", which = "cache")
 # load("./R/sysdata.rda",envir=the)
 #  loads tevents_defaults and ratingsmapmelt
 the$constants_fn <- paste0( the$defaultcachedir, "/avpf_constants.RD")
+the$NY_local_hrs = as.POSIXct(paste0(Sys.Date()," 12:00:00"),tz="UTC") - as.POSIXct(paste0(Sys.Date()," 12:00:00"),tz="America/New_York")
 the$pxd <- data.table()
 the$pxinv <- data.table()
+# Other item in the, downloaded on startup
+the$indexlist <- data.table()
 
 if(file.exists(the$constants_fn)) {
   restore_avs_state()